+++ test/fixtures/linter/indentation_errors.sql
@@ -1,6 +1,6 @@
 SELECT
    a.id, -- 3 Spaces
 	  a.name, -- tabs and spaces, 2 spaces
-	    a.training_spaces,   
+	    a.trailing_spaces,   
 	a.tabs_alone_after_spaces
 FROM tbl as a